add functions to utils.js - make esier to test
