+++ apps/src/p5lab/spritelab/blocks.js
@@ -9,6 +9,7 @@ import {Goal, showBackground} from '../redux/animationPicker';
 import i18n from '@cdo/locale';
 import spritelabMsg from '@cdo/spritelab/locale';
 import experiments from '@cdo/apps/util/experiments';
+import {parsePathString} from '@cdo/apps/p5lab/utils';
 
 function animations(includeBackgrounds) {
   const animationList = getStore().getState().animationList;
@@ -200,32 +201,6 @@ const customInputTypes = {
         dashboard.assets.showAssetManager(onSelect, 'audio', null, {
           libraryOnly: true,
         });
-      };
-      const capitalizeFirstLetter = function (string) {
-        return string.charAt(0).toUpperCase() + string.slice(1);
-      };
-      const parsePathString = text => {
-        // Example string paths:
-        // 'sound://category_board_games/card_dealing_multiple.mp3'
-        // 'sound://default.mp3'
-        const pathStringArray = text.split('/');
-        let category = '';
-        // Some sounds do not include a category, such as default.mp3
-        if (pathStringArray[2].includes('category_')) {
-          // Example: 'category_board_games' becomes 'Board games: '
-          category = capitalizeFirstLetter(
-            pathStringArray[2].replace('category_', '').replaceAll('_', ' ') +
-              ': '
-          );
-        }
-        // Example: 'card_dealing_multiple.mp3' becomes 'card_dealing_multiple'
-        const soundName = pathStringArray[pathStringArray.length - 1].replace(
-          '.mp3',
-          ''
-        );
-        // Examples: 'Board games: card_dealing_multiple', 'default'
-        const fieldText = `${category}${soundName}`;
-        return fieldText;
       };
       const onDisplay = soundPath => {
         return parsePathString(soundPath);
